pyproject.toml
[tool.poetry]
name = "map-kit-backend"
version = "0.1.0"
description = "The backend of the DCB Map Kit"
authors = ["Theo Costea <theo.costea@gmail.com>"]
license = "Apache License"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0"
djangorestframework = "^3.13"
psycopg2-binary = "^2.8"
djangorestframework-gis = "^1.0"
django-cors-headers = "^3.7"
httpx = "^0.23"
geocoder = "^1.38"
django-filter = "^2.4"
gunicorn = "^20.1"
whitenoise = "^5.2"
django-health-check = "^3.16.5"
python-jose = "^3.3.0"
social-auth-app-django = "^5.0.0"
djangorestframework-simplejwt = "^5.2.0"
drf-spectacular = "^0.22.1"

[tool.poetry.dev-dependencies]
black = "^22.3"
pytest = "^7.1.2"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "map_kit.settings"

[tool.black]
exclude = '(migrations)'
